ref: externalized UserProfile subcomponents
Externalized these subcomponents so that they are not server rendered. This is necessary so that the UserInformation component can be used in the `LoadingUserInformation` component that is used as the fallback while the real UserInformation component is loaded.
